protobuf size increased with  CodedInputStream.setSizeLimit()
